+++ src/main/java/com/example/companydbapi/repository/EmployeeReporitory.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.companydbapi.model.Employee;
 
-public interface EmployeeReporitory extends JpaRepository<Employee, Integer> {
+public interface EmployeeReporitory extends JpaRepository<Employee, Long> {
 }
